Optimized the search with before and after methods;
With the before and after method you can get your cursor just before a group or item of your search where the first item is >=search. With after it will be placed behind the block where the first element you get back is <=search.

Speedup the creation of sstable by not calling writeLong on a RandomAccessFile. That invokes 8 native write operations on the RandomAccessFile which slows down the serialization dramatically. Instead it uses an ByteBufferOutput stream to collect the bytes of a whole document before writing it to file.