Speedup the process of making indexes;
Indexes can now made partially of sstables and merged in a later stage. This saves a lot of time on bigger (which do not fit in memory) datasets. Every sstable is read in sequentially and indexed as is. This means that multiple writes to the same field are both indexed.

This can be solved by doing a compaction on your data (which eliminates multiple writes) or by changing the merge of index tables to check if the value encountered is the latest version of an entry, otherwise skip it.